Added mean standard deviation and clipping annotation vectors
